Separated out publish and consume actors so that they can be used independently. Made UnmanagedBaseActor use them and still be backward compatible
