+++ src/main/java/io/dropwizard/actors/actor/Actor.java
@@ -3,12 +3,11 @@ package io.dropwizard.actors.actor;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.dropwizard.actors.connectivity.RMQConnection;
 import io.dropwizard.actors.retry.RetryStrategyFactory;
+import java.util.Set;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
-
-import java.util.Set;
 
 /**
  * A simpler derivation of {@link BaseActor} to be used in most common actor use cases. This is managed by dropwizard.
@@ -17,21 +16,19 @@ import java.util.Set;
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)
 @Slf4j
-public abstract class  Actor<MessageType extends Enum<MessageType>, Message> extends BaseActor<Message> {
+public abstract class Actor<MessageType extends Enum<MessageType>, Message> extends BaseActor<Message> {
 
-    private MessageType type;
+  private MessageType type;
 
-    protected Actor(
-            MessageType type,
-            ActorConfig config,
-            RMQConnection connection,
-            ObjectMapper mapper,
-            RetryStrategyFactory retryStrategyFactory,
-            Class<? extends Message> clazz,
-            Set<Class<?>> droppedExceptionTypes) {
-        super(type.name(), config, connection, mapper, retryStrategyFactory, clazz, droppedExceptionTypes);
-        this.type = type;
-    }
-
-
+  protected Actor(
+      MessageType type,
+      ActorConfig config,
+      RMQConnection connection,
+      ObjectMapper mapper,
+      RetryStrategyFactory retryStrategyFactory,
+      Class<? extends Message> clazz,
+      Set<Class<?>> droppedExceptionTypes) {
+    super(type.name(), config, connection, mapper, retryStrategyFactory, clazz, droppedExceptionTypes);
+    this.type = type;
+  }
 }
