+++ src/main/java/io/dropwizard/actors/actor/UnmanagedConsumeActor.java
@@ -1,0 +1,126 @@
+package io.dropwizard.actors.actor;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.Lists;
+import com.rabbitmq.client.AMQP;
+import com.rabbitmq.client.Channel;
+import com.rabbitmq.client.DefaultConsumer;
+import com.rabbitmq.client.Envelope;
+import io.dropwizard.actors.connectivity.RMQConnection;
+import io.dropwizard.actors.retry.RetryStrategy;
+import io.dropwizard.actors.retry.RetryStrategyFactory;
+import io.dropwizard.actors.utils.NamingUtils;
+import java.io.IOException;
+import java.util.List;
+import java.util.function.Function;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class UnmanagedConsumeActor<Message> {
+
+  private final String name;
+  private final ActorConfig config;
+  private final RMQConnection connection;
+  private final ObjectMapper mapper;
+  private final Class<? extends Message> clazz;
+  private final int prefetchCount;
+  private final MessageHandlingFunction<Message, Boolean> handlerFunction;
+  private final Function<Throwable, Boolean> errorCheckFunction;
+  private final String queueName;
+  private final RetryStrategy retryStrategy;
+
+  private List<Handler> handlers = Lists.newArrayList();
+
+  public UnmanagedConsumeActor(
+      String name,
+      ActorConfig config,
+      RMQConnection connection,
+      ObjectMapper mapper,
+      RetryStrategyFactory retryStrategyFactory,
+      Class<? extends Message> clazz,
+      MessageHandlingFunction<Message, Boolean> handlerFunction,
+      Function<Throwable, Boolean> errorCheckFunction) {
+    this.name = name;
+    this.config = config;
+    this.connection = connection;
+    this.mapper = mapper;
+    this.clazz = clazz;
+    this.prefetchCount = config.getPrefetchCount();
+    this.handlerFunction = handlerFunction;
+    this.errorCheckFunction = errorCheckFunction;
+    this.queueName = NamingUtils.queueName(config.getPrefix(), name);
+    this.retryStrategy = retryStrategyFactory.create(config.getRetryConfig());
+  }
+
+  private boolean handle(Message message) throws Exception {
+    return handlerFunction.apply(message);
+  }
+
+  private class Handler extends DefaultConsumer {
+
+    private final ObjectMapper mapper;
+    private final Class<? extends Message> clazz;
+
+    @Getter
+    @Setter
+    private String tag;
+
+    private Handler(Channel channel,
+        ObjectMapper mapper,
+        Class<? extends Message> clazz,
+        int prefetchCount) throws Exception {
+      super(channel);
+      this.mapper = mapper;
+      this.clazz = clazz;
+      getChannel().basicQos(prefetchCount);
+    }
+
+    @Override
+    public void handleDelivery(String consumerTag, Envelope envelope,
+        AMQP.BasicProperties properties, byte[] body) throws IOException {
+      try {
+        final Message message = mapper.readValue(body, clazz);
+        boolean success = retryStrategy.execute(() -> handle(message));
+        if (success) {
+          getChannel().basicAck(envelope.getDeliveryTag(), false);
+        } else {
+          getChannel().basicReject(envelope.getDeliveryTag(), false);
+        }
+      } catch (Throwable t) {
+        log.error("Error processing message...", t);
+        if (errorCheckFunction.apply(t)) {
+          log.warn("Acked message due to exception: ", t);
+          getChannel().basicAck(envelope.getDeliveryTag(), false);
+        } else {
+          getChannel().basicReject(envelope.getDeliveryTag(), false);
+        }
+      }
+    }
+  }
+
+  public void start() throws Exception {
+    for (int i = 1; i <= config.getConcurrency(); i++) {
+      Channel consumeChannel = connection.newChannel();
+      final Handler handler = new Handler(consumeChannel, mapper, clazz, prefetchCount);
+      final String tag = consumeChannel.basicConsume(queueName, false, handler);
+      handler.setTag(tag);
+      handlers.add(handler);
+      log.info("Started consumer {} of type {}", i, name);
+    }
+  }
+
+  public void stop() throws Exception {
+    handlers.forEach(handler -> {
+      try {
+        final Channel channel = handler.getChannel();
+        channel.basicCancel(handler.getTag());
+        channel.close();
+      } catch (Exception e) {
+        log.error(String.format("Error cancelling consumer: %s", handler.getTag()), e);
+      }
+    });
+  }
+
+}
