+++ src/main/java/io/dropwizard/actors/utils/NamingUtils.java
@@ -1,0 +1,8 @@
+package io.dropwizard.actors.utils;
+
+public class NamingUtils {
+
+  public static String queueName(String prefix, String name) {
+    return String.format("%s.%s", prefix, name);
+  }
+}
